+++ pg-core/src/java/org/pg/reducer/Dummy.java
@@ -2,6 +2,7 @@ package org.pg.reducer;
 
 import org.pg.clojure.LazyMap;
 
+@SuppressWarnings("unused")
 public class Dummy implements IReducer {
 
     public static IReducer INSTANCE = new Dummy();