+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - pg.integration: config-txt and config-bin vars
 - pg.honey: on-connection wrapper
 - test json namespaces
+- rename Accum to Result
 
 ## 0.1.11